* Fix some 64-to-32-bit conversion warnings. Fixes #7459

* Fix typos in comments. Fixes #7761

* Fix error in example embedded in basic_socket::get_option's documentation. Fixes #7562

* Use long rather than int for SSL_CTX options, to match OpenSSL. Fixes #7209

* Use _snwprintf to address a compile error due to the changed swprintf signature in recent versions of MinGW. Fixes #7373

* Fix deadlock that can occur on Windows when shutting down a pool of io_service threads due to running out of work. Fixes #7552

* Enable noexcept qualifier for error categories. Fixes #7797

* Treat errors from accept as non-fatal. Fixes #7488

* Add a small block recycling optimisation.

* Version bump.

* Regenerate documentation.


[SVN r82290]

+++ include/boost/asio/ssl/detail/impl/engine.ipp
@@ -39,7 +39,8 @@ engine::engine(SSL_CTX* context)
 {
   if (!ssl_)
   {
-    boost::system::error_code ec(::ERR_get_error(),
+    boost::system::error_code ec(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     boost::asio::detail::throw_error(ec, "engine");
   }
@@ -166,7 +167,7 @@ boost::asio::mutable_buffers_1 engine::get_output(
 {
   int length = ::BIO_read(ext_bio_,
       boost::asio::buffer_cast<void*>(data),
-      boost::asio::buffer_size(data));
+      static_cast<int>(boost::asio::buffer_size(data)));
 
   return boost::asio::buffer(data,
       length > 0 ? static_cast<std::size_t>(length) : 0);
@@ -177,7 +178,7 @@ boost::asio::const_buffer engine::put_input(
 {
   int length = ::BIO_write(ext_bio_,
       boost::asio::buffer_cast<const void*>(data),
-      boost::asio::buffer_size(data));
+      static_cast<int>(boost::asio::buffer_size(data)));
 
   return boost::asio::buffer(data +
       (length > 0 ? static_cast<std::size_t>(length) : 0));
@@ -228,7 +229,7 @@ engine::want engine::perform(int (engine::* op)(void*, std::size_t),
   std::size_t pending_output_before = ::BIO_ctrl_pending(ext_bio_);
   int result = (this->*op)(data, length);
   int ssl_error = ::SSL_get_error(ssl_, result);
-  int sys_error = ::ERR_get_error();
+  int sys_error = static_cast<int>(::ERR_get_error());
   std::size_t pending_output_after = ::BIO_ctrl_pending(ext_bio_);
 
   if (ssl_error == SSL_ERROR_SSL)
@@ -296,12 +297,14 @@ int engine::do_shutdown(void*, std::size_t)
 
 int engine::do_read(void* data, std::size_t length)
 {
-  return ::SSL_read(ssl_, data, length < INT_MAX ? length : INT_MAX);
+  return ::SSL_read(ssl_, data,
+      length < INT_MAX ? static_cast<int>(length) : INT_MAX);
 }
 
 int engine::do_write(void* data, std::size_t length)
 {
-  return ::SSL_write(ssl_, data, length < INT_MAX ? length : INT_MAX);
+  return ::SSL_write(ssl_, data,
+      length < INT_MAX ? static_cast<int>(length) : INT_MAX);
 }
 
 #endif // !defined(BOOST_ASIO_ENABLE_OLD_SSL)
