+++ include/boost/asio/ssl/impl/context.ipp
@@ -91,7 +91,8 @@ context::context(context::method m)
 
   if (handle_ == 0)
   {
-    boost::system::error_code ec(::ERR_get_error(),
+    boost::system::error_code ec(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     boost::asio::detail::throw_error(ec, "context");
   }
@@ -201,7 +202,8 @@ boost::system::error_code context::load_verify_file(
 {
   if (::SSL_CTX_load_verify_locations(handle_, filename.c_str(), 0) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -222,7 +224,8 @@ boost::system::error_code context::set_default_verify_paths(
 {
   if (::SSL_CTX_set_default_verify_paths(handle_) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -243,7 +246,8 @@ boost::system::error_code context::add_verify_path(
 {
   if (::SSL_CTX_load_verify_locations(handle_, 0, path.c_str()) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -282,7 +286,8 @@ boost::system::error_code context::use_certificate_file(
 
   if (::SSL_CTX_use_certificate_file(handle_, filename.c_str(), file_type) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -303,7 +308,8 @@ boost::system::error_code context::use_certificate_chain_file(
 {
   if (::SSL_CTX_use_certificate_chain_file(handle_, filename.c_str()) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -342,7 +348,8 @@ boost::system::error_code context::use_private_key_file(
 
   if (::SSL_CTX_use_PrivateKey_file(handle_, filename.c_str(), file_type) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -382,7 +389,8 @@ boost::system::error_code context::use_rsa_private_key_file(
   if (::SSL_CTX_use_RSAPrivateKey_file(
         handle_, filename.c_str(), file_type) != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -417,11 +425,12 @@ boost::system::error_code context::use_tmp_dh_file(
   }
 
   ::BIO_free(bio);
-  int result = ::SSL_CTX_set_tmp_dh(handle_, dh);
+  long result = ::SSL_CTX_set_tmp_dh(handle_, dh);
   ::DH_free(dh);
   if (result != 1)
   {
-    ec = boost::system::error_code(::ERR_get_error(),
+    ec = boost::system::error_code(
+        static_cast<int>(::ERR_get_error()),
         boost::asio::error::get_ssl_category());
     return ec;
   }
@@ -510,7 +519,7 @@ int context::password_callback_function(
     strncat(buf, passwd.c_str(), size);
 #endif
 
-    return strlen(buf);
+    return static_cast<int>(strlen(buf));
   }
 
   return 0;
